Extract block confirmation to its own files
